# doc/api.yml
openapi: 3.0.3
info:
  version: '1.0'
  title: Voting.com
  description: voting api with all the verbs 

servers:
  - url: 'https://api-dev.hei.school'
  - url: 'https://api-prod.hei.school'

security:
  - BearerAuth: [ ]

paths:
  '/auth/login':
    post:
      tags:
        - Security
      summary: Tells you who you are.
      requestBody:
        required: true
        $ref: '#/components/requestBodies/postBody'
      operationId: Login
      responses:
        '200':
          description: Who you are
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/whoami'
        '401':
          $ref: '#/components/responses/401'
  '/auth/signup':
    get:
      tags:
        - Security
      summary: Tells you who you are after subscribing.
      operationId: signup
      responses:
        '200':
          description: Who you are
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Signup'
        '401':
          $ref: '#/components/responses/401'
  '/sondage':
    get:
      tags:
      - Sondage
      summary: list of the polls
      operationId: getSondage
      responses:
        '200':
          description: idk
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sondage'
        '401':
          $ref: '#/components/responses/401'

  '/sondage/{id}':
    get:
      tags:
        - Sondage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      summary: the specific poll that will be answered
      operationId: getSondageId
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/Sondage'
        '401':
          $ref: '#/components/responses/401'
    put:
      tags:
        - Sondage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      summary: Modify the answers
      requestBody:
        required: true
        $ref: '#/components/requestBodies/postSondageBody'
      operationId: putSondageId
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSondage'
        '401':
          $ref: '#/components/responses/401'

components:
  requestBodies:
    postBody:
      description: a JSON object containing login informations
      required: true
      content:
        application/json:
          schema:
            $ref: 
              '#/components/schemas/Login'
    postSondageBody:
      description: a JSON object containing the answers
      required: true
      content:
        application/json:
          schema:
            $ref:
              '#/components/schemas/Sondage'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer

  responses:
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotAuthorizedException'
  schemas:
    whoami:
      type: object
      properties:
        email:
          type: string
        bearer:
          type: string
    Login:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    Signup:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
    Sondage:
      type: object
      properties:
        id: 
          type: integer
        title:
          type: string
        description:
          type: string
    CreateSondage:
      type: object
      properties:
        id: 
          type: integer
        title:
          type: string
        description:
          type: string
        questions:
          type: array
          items:
            type: object
            $ref: '#/components/schemas/Question'
    Question: 
      type: object
      properties:
        sentence:
          type: string
        type:
          type: string
        options:
          type: array
          items:
            type: string
    User:
      type: object
      properties:
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - USER

    Page:
      type: integer
      description: Start at `1`
    PageSize:
      type: integer
      description: Cannot exceed `500`

    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: NotAuthorizedException
        message: Not authorized